move old scripts
